update: sdxl integration (remove edge case)
