+++ tests/functional_tests/t0_main/diffusers/t1_sdxl_generation.yea
@@ -22,7 +22,7 @@ assert:
     - :wandb:runs[0][config][workflow][0][params][output_type]: "latent"
     - :wandb:runs[0][config][workflow][0][params][prompt]: "A small cactus with a happy face in the Sahara desert."
     - :wandb:runs[0][config][workflow][0][params][generator][seed]: 10
-    - :wandb:runs[0][config][workflow][0][pipeline][_name_or_path]: "stabilityai/stable-diffusion-xl-base-1.0"
+    - :wandb:runs[0][config][workflow][0][pipeline][_name_or_path]: "stabilityai/stable-diffusion-xl-Pipeline-Call-1-1.0"
     - :wandb:runs[0][config][workflow][0][pipeline][force_zeros_for_empty_prompt]: true
     - :wandb:runs[0][config][workflow][0][pipeline][pipeline-name]: "StableDiffusionXLPipeline"
     - :wandb:runs[0][config][workflow][0][pipeline][scheduler][0]: "diffusers"
@@ -39,7 +39,7 @@ assert:
     - :wandb:runs[0][config][workflow][0][pipeline][unet][1]: "UNet2DConditionModel"
     - :wandb:runs[0][config][workflow][0][pipeline][vae][0]: "diffusers"
     - :wandb:runs[0][config][workflow][0][pipeline][vae][1]: "AutoencoderKL"
-    - :wandb:runs[0][config][workflow][0][stage]: "Base"
+    - :wandb:runs[0][config][workflow][0][stage]: "Pipeline-Call-1"
     - :wandb:runs[0][config][workflow][1][params][aesthetic_score]: 6
     - :wandb:runs[0][config][workflow][1][params][eta]: 0
     - :wandb:runs[0][config][workflow][1][params][guidance_rescale]: 0
@@ -53,7 +53,7 @@ assert:
     - :wandb:runs[0][config][workflow][1][params][prompt]: "A small cactus with a happy face in the Sahara desert."
     - :wandb:runs[0][config][workflow][1][params][generator][seed]: 10
     - :wandb:runs[0][config][workflow][1][params][strength]: 0.3
-    - :wandb:runs[0][config][workflow][1][pipeline][_name_or_path]: "stabilityai/stable-diffusion-xl-refiner-1.0"
+    - :wandb:runs[0][config][workflow][1][pipeline][_name_or_path]: "stabilityai/stable-diffusion-xl-Pipeline-Call-2-1.0"
     - :wandb:runs[0][config][workflow][1][pipeline][force_zeros_for_empty_prompt]: false
     - :wandb:runs[0][config][workflow][1][pipeline][pipeline-name][pipeline-name]: "StableDiffusionXLImg2ImgPipeline"
     - :wandb:runs[0][config][workflow][1][pipeline][pipeline-name][requires_aesthetics_score]: true
@@ -63,19 +63,19 @@ assert:
     - :wandb:runs[0][config][workflow][1][pipeline][unet][1]: "UNet2DConditionModel"
     - :wandb:runs[0][config][workflow][1][pipeline][vae][0]: "diffusers"
     - :wandb:runs[0][config][workflow][1][pipeline][vae][1]: "AutoencoderKL"
-    - :wandb:runs[0][config][workflow][1][pipeline][stage]: "Refiner"
-    - :wandb:runs[0][summary][Base][Result-Table][_type]: table-file
-    - :wandb:runs[0][summary][Base][Result-Table][nrows]: 1
-    - :wandb:runs[0][summary][Refiner][Result-Table][_type]: table-file
-    - :wandb:runs[0][summary][Refiner][Result-Table][nrows]: 1
-    - :wandb:runs[0][summary][Base][Generated-Image][_type]: image-file
-    - :wandb:runs[0][summary][Base][Generated-Image][caption]: "Prompt-1: A small cactus with a happy face in the Sahara desert.\nPrompt-2: None"
-    - :wandb:runs[0][summary][Base][Generated-Image][format]: "png"
-    - :wandb:runs[0][summary][Base][Generated-Image][height]: 1024
-    - :wandb:runs[0][summary][Base][Generated-Image][width]: 1024
-    - :wandb:runs[0][summary][Refiner][Generated-Image][_type]: image-file
-    - :wandb:runs[0][summary][Refiner][Generated-Image][caption]: "Prompt-1: A small cactus with a happy face in the Sahara desert.\nPrompt-2: None"
-    - :wandb:runs[0][summary][Refiner][Generated-Image][format]: "png"
-    - :wandb:runs[0][summary][Refiner][Generated-Image][height]: 1024
-    - :wandb:runs[0][summary][Refiner][Generated-Image][width]: 1024
+    - :wandb:runs[0][config][workflow][1][pipeline][stage]: "Pipeline-Call-2"
+    - :wandb:runs[0][summary][Result-Table][Pipeline-Call-1][_type]: table-file
+    - :wandb:runs[0][summary][Result-Table][Pipeline-Call-1][nrows]: 1
+    - :wandb:runs[0][summary][Result-Table][Pipeline-Call-2][_type]: table-file
+    - :wandb:runs[0][summary][Result-Table][Pipeline-Call-2][nrows]: 1
+    - :wandb:runs[0][summary][Generated-Image][Pipeline-Call-1][_type]: image-file
+    - :wandb:runs[0][summary][Generated-Image][Pipeline-Call-1][caption]: "Prompt-1: A small cactus with a happy face in the Sahara desert.\nPrompt-2: None"
+    - :wandb:runs[0][summary][Generated-Image][Pipeline-Call-1][format]: "png"
+    - :wandb:runs[0][summary][Generated-Image][Pipeline-Call-1][height]: 1024
+    - :wandb:runs[0][summary][Generated-Image][Pipeline-Call-1][width]: 1024
+    - :wandb:runs[0][summary][Generated-Image][Pipeline-Call-1][_type]: image-file
+    - :wandb:runs[0][summary][Generated-Image][Pipeline-Call-1][caption]: "Prompt-1: A small cactus with a happy face in the Sahara desert.\nPrompt-2: None"
+    - :wandb:runs[0][summary][Generated-Image][Pipeline-Call-1][format]: "png"
+    - :wandb:runs[0][summary][Generated-Image][Pipeline-Call-1][height]: 1024
+    - :wandb:runs[0][summary][Generated-Image][Pipeline-Call-1][width]: 1024
     - :wandb:runs[0][exitcode]: 0
