+++ wandb/integration/diffusers/resolvers/__init__.py
@@ -2,11 +2,8 @@ from .multimodal import (
     SUPPORTED_MULTIMODAL_PIPELINES,
     DiffusersMultiModalPipelineResolver,
 )
-from .sdxl import SUPPORTED_SDXL_PIPELINES, SDXLResolver
 
 __all__ = [
     "SUPPORTED_MULTIMODAL_PIPELINES",
-    "SUPPORTED_SDXL_PIPELINES",
     "DiffusersMultiModalPipelineResolver",
-    "SDXLResolver",
 ]
