+++ wandb/integration/diffusers/resolvers/utils.py
@@ -6,6 +6,7 @@ from wandb.util import get_module
 
 if TYPE_CHECKING:
     np_array = get_module("numpy.array")
+    torch_float_tensor = get_module("torch.FloatTensor")
 
 
 def chunkify(input_list, chunk_size) -> List:
@@ -63,3 +64,39 @@ def postprocess_np_arrays_for_video(
     )
     images = [(img * 255).astype("uint8") for img in images] if normalize else images
     return np.transpose(np.stack((images), axis=0), axes=(0, 3, 1, 2))
+
+
+def decode_sdxl_t2i_latents(pipeline: Any, latents: "torch_float_tensor") -> List:
+    """Decode latents generated by [`diffusers.StableDiffusionXLPipeline`](https://huggingface.co/docs/diffusers/main/en/api/pipelines/stable_diffusion/stable_diffusion_xl#stable-diffusion-xl).
+
+    Arguments:
+        pipeline: (diffusers.DiffusionPipeline) The Diffusion Pipeline from
+            [`diffusers`](https://huggingface.co/docs/diffusers).
+        latents (torch.FloatTensor): The generated latents.
+
+    Returns:
+        List of `PIL` images corresponding to the generated latents.
+    """
+    torch = get_module(
+        "torch",
+        required="Please ensure PyTorch is installed. You can check out https://pytorch.org/get-started/locally/#start-locally for installation instructions.",
+    )
+    with torch.no_grad():
+        needs_upcasting = (
+            pipeline.vae.dtype == torch.float16 and pipeline.vae.config.force_upcast
+        )
+        if needs_upcasting:
+            pipeline.upcast_vae()
+            latents = latents.to(
+                next(iter(pipeline.vae.post_quant_conv.parameters())).dtype
+            )
+        images = pipeline.vae.decode(
+            latents / pipeline.vae.config.scaling_factor, return_dict=False
+        )[0]
+        if needs_upcasting:
+            pipeline.vae.to(dtype=torch.float16)
+        if pipeline.watermark is not None:
+            images = pipeline.watermark.apply_watermark(images)
+        images = pipeline.image_processor.postprocess(images, output_type="pil")
+        pipeline.maybe_free_model_hooks()
+        return images
